third_party/WebKit/LayoutTests/bluetooth/requestDevice/discovery-succeeds.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/testdriver.js"></script>
<script src="../../resources/testdriver-vendor.js"></script>
<script src="../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
const test_desc = 'Discover a device using alias, name, or UUID.';

bluetooth_test(() => getHealthThermometerDevice()
    .then(() => requestDeviceWithTrustedClick({
      filters: [{ services: [health_thermometer.alias] }]
    }))
    .then(device => assert_equals(device.constructor.name, 'BluetoothDevice'))
    .then(() => requestDeviceWithTrustedClick({
      filters: [{ services: [health_thermometer.name] }]
    }))
    .then(device => assert_equals(device.constructor.name, 'BluetoothDevice'))
    .then(() => requestDeviceWithTrustedClick({
      filters: [{ services: [health_thermometer.uuid] }]
    }))
    .then(device => assert_equals(device.constructor.name, 'BluetoothDevice')),
test_desc);
</script>
